results number beneath event search Search button now displays filtered results number. Moves filtered events logic from scenes service to sarviews event service and changes references accordingly
